requirements.txt
uvicorn
fastapi
langchain_groq
langchain_huggingface
python-dotenv
pydantic
python-multipart
pandas
numpy
matplotlib
seaborn
pymongo
cloudinary